requirements.txt
asyncio==3.4.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
ffmpeg-python==0.2.0
future==0.18.2
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
functions-framework
idna==2.10
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==7.0.2
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
urllib3==1.25.10

